requirements.txt
-e .
wheel
pytest
aws_cdk.aws_cloudformation
aws_cdk.aws_cloudtrail
aws_cdk.aws_ec2
aws_cdk.aws_ecs
aws_cdk.aws_ecr
aws-cdk.aws_iam
aws_cdk.aws_s3
aws_cdk.aws_s3_deployment
